requirements.txt
streamlit
youtube_transcript_api
together
gTTS
python-dotenv 